+++ R/MultipleSetupAndPosterior.R
@@ -10,7 +10,7 @@
 #'
 #' @param checkpoint.chlist If \code{TRUE}, checkpoint the (unclean)
 #'    chlist to "initial.chlist.Rdata" in the current working directory.
-#'    and checkpoint the clean clist ot "clean.chlist.Rdata" in the
+#'    and checkpoint the clean chlist to "clean.chlist.Rdata" in the
 #'    current working directory.
 #'
 #' @return Invisibly,